gts210wifi: add LOCAL_MODULE_STEM to Android.mk for ueventd.rc
gts28wifi had this in it's Androik.mk and worked with 19.1.  Without
this gts210wifi displays an error message

mali     ERROR: The DDK is not compatible with any of the Mali GPUs on the system.
mali     The DDK was built for 0x750 r0p1 status range [0..15], but none of the GPUs matched:

even though the same libGLES_mali.so worked with LineageOS 17.1, 18.1
and 19.1 gts28wifi and 17.1 and 18.1 gts210wifi.

The kernel correctly identifies the GPU as

[c7] mali 14ac0000.mali: GPU identified as 0x0750 r0p1 status 1

More documentation at

https://android.googlesource.com/platform/build/+/master/Changes.md

Removing ‘/’ from Valid Module Names {#name_slash}
The build system uses module names in path names in many places. Having an extra ‘/’ or ‘../’ being inserted can cause problems -- and not just build breaks, but stranger invalid behavior.

In every case we‘ve seen, the fix is relatively simple: move the directory into LOCAL_MODULE_RELATIVE_PATH (or LOCAL_MODULE_PATH if you’re still using it). If this causes multiple modules to be named the same, use unique module names and LOCAL_MODULE_STEM to change the installed file name: